+++ .gitignore
@@ -18,3 +18,4 @@ tested.code
 .sass-cache
 github_metrics
 fileformaterrors
+localmetricsstore
